טסטים מלאים ותיקונים

+++ src/DBDAO/CompanyDBDAO.java
@@ -98,34 +98,45 @@ public class CompanyDBDAO implements CompanyDAO {
 			
 			long companyId = company.getId();
 			CouponDAO couponDAO = new CouponDBDAO();
-			List<Coupon> couponsOwnedByCompany = couponDAO.getAllCouponsCompany(companyId);
+			
+			List<Coupon> couponsOwnedByCompany = null;
+			boolean flag = true;
+			
+			try{
+				couponsOwnedByCompany = couponDAO.getAllCouponsCompany(companyId);
+			}catch(CouponSystemException e){
+				flag = false;
+			}
 			
 			connection.setAutoCommit(false);
 			
-			for (Coupon coupon : couponsOwnedByCompany) {
-				
-				long couponId = coupon.getId();
-				sql = "DELETE FROM customer_coupon WHERE coupon_id = ?";
-				preparedSt = connection.prepareStatement(sql);
-				
-				preparedSt.setLong(1, couponId);
-				preparedSt.executeUpdate();
-				preparedSt.close();
-				
-				sql = "DELETE FROM company_coupon WHERE coupon_id = ?";
-				preparedSt = connection.prepareStatement(sql);
-				
-				preparedSt.setLong(1, couponId);
-				preparedSt.executeUpdate();
-				preparedSt.close();
-				
-				sql = "DELETE FROM coupon WHERE id = ?";
-				preparedSt = connection.prepareStatement(sql);
-				
-				preparedSt.setLong(1, couponId);
-				preparedSt.executeUpdate();
-				preparedSt.close();
+			if(flag){
+				for (Coupon coupon : couponsOwnedByCompany) {
+					
+					long couponId = coupon.getId();
+					sql = "DELETE FROM customer_coupon WHERE coupon_id = ?";
+					preparedSt = connection.prepareStatement(sql);
+					
+					preparedSt.setLong(1, couponId);
+					preparedSt.executeUpdate();
+					preparedSt.close();
+					
+					sql = "DELETE FROM company_coupon WHERE coupon_id = ?";
+					preparedSt = connection.prepareStatement(sql);
+					
+					preparedSt.setLong(1, couponId);
+					preparedSt.executeUpdate();
+					preparedSt.close();
+					
+					sql = "DELETE FROM coupon WHERE id = ?";
+					preparedSt = connection.prepareStatement(sql);
+					
+					preparedSt.setLong(1, couponId);
+					preparedSt.executeUpdate();
+					preparedSt.close();
+				}
 			}
+			
 			
 			sql = "DELETE FROM company WHERE id = ?";
 			preparedSt = connection.prepareStatement(sql);
@@ -190,7 +201,7 @@ public class CompanyDBDAO implements CompanyDAO {
 		
 		try {
 			
-			sql = "SELECT * FROM company WHERE company.id = '" + id + "' FETCH FIRST ROW ONLY";
+			sql = "SELECT * FROM company WHERE company.id = " + id + "FETCH FIRST ROW ONLY";
 			statement = connection.createStatement();
 			result = statement.executeQuery(sql);
 			if (result.next()) {
